% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{change.count}
\alias{change.count}
\title{Zmiana liczby danego zwierzęcia w stadzie}
\usage{
change.count(game, animal, change)
}
\arguments{
\item{game}{Tabela gry.}

\item{animal}{Nazwa zwierzęcia którego stan będziemy zwiększać.}

\item{change}{Wartość o którą zwiększamy liczebność danego zwierzęcia.}
}
\description{
Funkcja \code{change.count} zmienia liczbę danego zwierzęcia w stadzie o wskazaną wartość.
}
\examples{
change.count(game,"rabbit",10)
change.count(game,"big_dog",1)

}

