% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{clear.count}
\alias{clear.count}
\title{Zmiana stanu danego zwierzęcia na zerowy}
\usage{
clear.count(game, animal)
}
\arguments{
\item{game}{Tabela gry SuperFarmer.}

\item{animal}{Nazwa zwierzęcia.}
}
\description{
Funkcja \code{clear.count}  zeruje liczbę zwierząt danego gatunku.
}
\examples{
clear.count(game,"horse")
clear.count(game,"pig")
clear.count(game,"rabbit")

}

