% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{convert.farm.vector}
\alias{convert.farm.vector}
\title{Przetwarzanie wektora do tablicy game}
\usage{
convert.farm.vector(farm)
}
\arguments{
\item{farm}{Wektor zawierający stan stada w chwili obecnej.}
}
\description{
Funkcja \code{convert.farm.vector} konwertuje wektor do tabeli game.
Służy ona do umożliwienia uruchomienia innych  strategii w naszym pakiecie.
}
\details{
Zgodnie z ustaleniami wektor, który ma być przekazywany jest innej postaci niż ustalony.
Nasza tabela game ma zwierzęta w kolejności królik, mały pies, owca, świnia, duży pies, krowa, koń.
W związku z tym konieczna jest funkcja przekazująca wektor do tablicy game.
}
\examples{
convert.farm.vector(c(7:1))
convert.farm.vector(c(10,2,1,1,0,0,0))

}

