% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{convert.game.table}
\alias{convert.game.table}
\title{Przetwarzanie tablicy game do wektora}
\usage{
convert.game.table(game)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}
}
\description{
Funkcja \code{convert.game.table} konwertuje tabelę game do wektora farm.
Służy ona do umożliwienia uruchomienia naszej strategii w innych pakietach gry.
}
\details{
Zgodnie z ustaleniami wektor, który ma być przekazywany jest innej postaci niż ustalony.
Nasza tabela game ma zwierzęta w kolejności królik, mały pies, owca, świnia, duży pies, krowa, koń.
W związku z tym konieczna jest funkcja konwertująca.
}
\examples{
convert.game.table(game)

}

