% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generuj_tabele_gry.R
\name{create.game.table}
\alias{create.game.table}
\title{Utworzenie tabeli game}
\usage{
create.game.table()
}
\value{
Tabela game, na której będą wykonywane strategie gry SuperFarmer.
}
\description{
Funkcja \code{create.game.table} tworzy pustą tabelę gry
}
\examples{
create.game.table()

}

