% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{examine.game}
\alias{examine.game}
\title{Badanie gry SuperFarmer}
\usage{
examine.game(rounds, namestrategy)
}
\arguments{
\item{rounds}{Liczba wykonanych powtórzeń danej gry.}

\item{namestrategy}{Nazwa wykonywanej strategii.}
}
\description{
Funkcja \code{examine.game} bada rozkład czasu gry SuperFarmer.
}
\examples{
examine.game(100,strategy.PDMS)

}

