% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_PDMS.R
\name{exchange.animals.for.rabbits}
\alias{exchange.animals.for.rabbits}
\title{Zamiana zwierząt na króliki}
\usage{
exchange.animals.for.rabbits(game, farm = c("small_dog", "sheep", "pig",
  "big_dog", "cow"))
}
\arguments{
\item{game}{Tabela gry SuperFarmer.}

\item{farm}{Wektor z podzbiorem stada.}
}
\value{
Tablica game po wymianie.
}
\description{
Funkcja \code{exchange.animals.for.rabbits} zamienia zwięrzęta na króliki.
W przypadku gdy mamy więcej niż jedno zwierze innego gatunku niż króliki zamienia "nadmiarowe" zwierzęta na króliki.
W przypadku gdy nie mamy królików: bierze zwierzę o największej wartości (poza koniem) jakie posiadamy i zamienia je na króliki.
}
\examples{
exchange.animals.for.rabbits(game)
exchange.animals.for.rabbits(game, farm = c("sheep","pig"))

}

