% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{exchange.farm}
\alias{exchange.farm}
\title{Zmiana liczebności podzbioru stada o wskazaną wartość}
\usage{
exchange.farm(game, farm, count)
}
\arguments{
\item{game}{Tabela gry zawierająca informacje dotyczące gry.}

\item{farm}{Wskazany podzbiór zwierząt, których liczebność chcemy zmieniać o wskazaną wartość.}

\item{count}{Wartość o którą zmieniamy.}
}
\description{
Funkcja \code{exchange.farm}zmienia liczebność wskazanego podzbioru stada
}
\examples{
exchange.farm(game,farm = c("rabbit","sheep","pig"),1)
exchange.farm(game,"big_dog",1)

}

