% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_PDMS.R
\name{exchange.rabbits.for.animals}
\alias{exchange.rabbits.for.animals}
\title{Zamiana królików na inne zwierzęta}
\usage{
exchange.rabbits.for.animals(game, farm = c("small_dog", "sheep", "pig",
  "big_dog", "cow"))
}
\arguments{
\item{game}{Tabela gry SuperFarmer.}

\item{farm}{Wektor z podzbiorem stada.}
}
\value{
Tablica game po wymianie.
}
\description{
Funkcja \code{exchange.rabbits.for.animals} zamienia króliki na zwierzęta.
}
\details{
Działanie: Jeśli nie mamy zwierzęcia a możemy go otrzymać z królików to zamieniamy.
}
\examples{
exchange.animals.for.rabbits(game)
exchange.animals.for.rabbits(game, farm = c("sheep","pig"))

}

