% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_fazy_rzutu_kostka.R
\name{fox.reaction}
\alias{fox.reaction}
\title{Reakcja na lisa}
\usage{
fox.reaction(game, die2)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}

\item{die2}{Zwierzę otrzymane w wyniku losowania na kostce nr 2.}
}
\value{
Tablica game po wykonaniu rzutu kostką.
}
\description{
Funkcja \code{fox.reaction} symuluje zmianę stada po wylosowaniu lisa.
}
\examples{
fox.reaction(game,"rabbit")
fox.reaction(game,"small_dog")

}

