% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{game.finished}
\alias{game.finished}
\title{Warunek kończący grę}
\usage{
game.finished(game)
}
\arguments{
\item{game}{Tabela gry SuperFarmer.}
}
\description{
Funkcja \code{game.finished} czy warunek końca rozgrywki jest spełniony.
}
\examples{
game.finished(game)

}

