% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{get.count}
\alias{get.count}
\title{Stan obecny danego zwierzęcia}
\usage{
get.count(game, animal)
}
\arguments{
\item{game}{Tabela gry.}

\item{animal}{Nazwa zwierzęcia.}
}
\description{
Funkcja \code{get.count} zwraca liczbę zwierząt danego gatunku, posiadanych w stadzie.
}
\examples{
get.count(game,"horse")
get.count(game,"pig")
get.count(game,"rabbit")

}

