% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{get.value}
\alias{get.value}
\title{Wartość danego zwierzęcia liczona w królikach}
\usage{
get.value(game, animal)
}
\arguments{
\item{game}{Tabela gry.}

\item{animal}{Nazwa zwierzęcia.}
}
\description{
Funkcja \code{get.value}  zwraca wartość danego zwierzęcia liczoną w królikach.
}
\examples{
get.value(game,"horse")
get.value(game,"pig")
get.value(game,"rabbit")

}

