% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_symulujace_gre.R
\name{make.move}
\alias{make.move}
\title{Pojedyńczy ruch w grze SuperFarmer}
\usage{
make.move(game, strategy)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}

\item{strategy}{Nazwa wykonywanej strategii.}
}
\description{
Funkcja \code{make.move} wykonuje jeden ruch w grze SuperFarmer.
Jako wartość zwraca tabelę game po wykonaniu jednego ruchu w grze.
}
\examples{
make.move(game,strategia_DKA)

}

