% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_symulujace_gre.R
\name{play}
\alias{play}
\title{Jednokrotne wykonanie symulacji  gry SuperFarmer}
\usage{
play(strategyname)
}
\arguments{
\item{strategyname}{Nazwa wykonywanej strategii.}
}
\description{
Funkcja \code{play} symuluje całą grę jeden raz.
Jako wartość zwraca liczbę rund potrzebnych do ukończenia gry.
}
\examples{
play(strategy.PDMS)

}

