% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{strategia.DKA}
\alias{strategia.DKA}
\title{Strategia DKA}
\usage{
strategia.DKA(stan_zagrody)
}
\arguments{
\item{stan_zagrody}{Wektor zawierający informację dotyczącego posiadanego stada w chwili obecnej.}
}
\value{
Wektor liczebności stada po wykonaniu strategi.
}
\description{
Funkcja \code{strategy.test}
}
\details{
Strategia jest strategią jeden  na wiele..
\itemize{
 \item Konwersji wektora do tabeli gra - możliwość uruchomienia  strategii zewnętrznych w naszym pakiecie.
 \item Dopisać co robi
 \item Konwersji do wektora - możliwość uruchomienia w innych pakietach.
}
}
\examples{
strategia.DKA(c(8,1,3,0,0,0,1))

}
\author{
Dominik Ambroziak, dopisz autorów z 1 fazy
}

