% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{strategia_DKA}
\alias{strategia_DKA}
\title{Strategia DKA}
\usage{
strategia_DKA(stan_zagrody)
}
\arguments{
\item{stan_zagrody}{Wektor zawierający informację dotyczącego posiadanego stada w chwili obecnej.}
}
\value{
Wektor liczebności stada po wykonaniu strategi.
}
\description{
Funkcja chceck wyrzuca 1 NOTE (niby to nie jest błąd, ale jak by łatwo to można poprawić)
wymiana_na_drozsze: no visible binding for global variable 'zwierzeta'
Warto to poprawić. W moim przypadku było konieczne coś takiego zrobić w przypadku kostek.
Dodałem więc kostki jako parametr do każdej funkcji likwidując w ten sposób zmienną globalną.
}
\details{
Funkcja \code{strategia_DKA}

Strategia jest strategią jeden  na wiele..
\itemize{
 \item Konwersji wektora do tabeli gra - możliwość uruchomienia  strategii zewnętrznych w naszym pakiecie.
 \item Dopisać co robi
 \item Konwersji do wektora - możliwość uruchomienia w innych pakietach.
}
}
\examples{
strategia_DKA(c(8,1,3,0,0,0,1))

}
\author{
Dominik Ambroziak, dopisz autorów z 1 fazy
}

