% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{strategy_DKA}
\alias{strategy_DKA}
\title{Strategia DKA}
\usage{
strategy_DKA(stan_zagrody)
}
\arguments{
\item{stan_zagrody}{Wektor zawierający informacje dotyczące posiadanego stada w chwili obecnej.}
}
\value{
Wektor liczebności stada po wykonaniu wymiany zgodnie z założeniami strategi.
}
\description{
Funkcja \code{strategy_DKA}
}
\details{
Strategia używana w pierwszym etapie przez grupę, w której skład wchodzili:
Dominik Ambroziak, Anna Dymowska, Adam Król. Podstawowe założenia startegii:
\itemize{
 \item Gdy jesteśmy w stanie wymienić tańsze zwierzęta na droższe - wymieniamy
 \item W drugą stronę wymieniamy tylko wtedy gdy jest to ruch kończący grę.
 \item Dużego psa kupujemy przed krową.
 \item Nie kupujemy małych psów.
 \item Staramy się zawsze mieć przynajmniej jednego królika. Nie dokonujemy
 wymian, które powodowałyby wyzerowanie stanu posiadanych królików.
}

Czynności wykonywane przez funkcję strategy_DKA:
\itemize{
 \item Konwersji wektora do tabeli "zwierzeta"
 \item Jeśli posiadane zwierzęta mają wartość wystarczającą do wygrania,
  a nie mamy jeszcze każdego z wymaganych zwierząt to wymieniamy tak, by mieć już wszystkie
 \item W przeciwnym przypadku dokonujemy zamiany tańszych zwierząt na droższe, w miarę możliwości.
 \item Konwersji do wektora - możliwość uruchomienia w innych pakietach.
}
}
\examples{
strategy_DKA(c(8,1,3,0,0,0,1))

}
\author{
Dominik Ambroziak, Anna Dymowska, Adam Król
}

