% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_fazy_rzutu_kostka.R
\name{throw.dice}
\alias{throw.dice}
\title{Faza rzutu kostką}
\usage{
throw.dice(game, die1, die2)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}

\item{die1}{Kostka nr1.}

\item{die2}{Kostka nr2.}
}
\value{
Tablica game po wykonaniu rzutu kostką.
}
\description{
Funkcja \code{throw.dice} symuluje fazę gry związaną z rzutem kostką.
}
\examples{
throw.dice(game,die1,die2)

}

