% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_fazy_rzutu_kostka.R
\name{wolf.reaction}
\alias{wolf.reaction}
\title{Reakcja na wilka}
\usage{
wolf.reaction(game, die1)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}

\item{die1}{Zwierzę otrzymane w wyniku losowania na kostce nr 1.}
}
\value{
Tablica game po wykonaniu rzutu kostką.
}
\description{
Funkcja \code{wolf.reaction} symuluje zmianę stada po wylosowaniu wilka.
}
\examples{
wolf.reaction(game,"rabbit")
wolf.reaction(game,"wolf")

}

