% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{examine.game}
\alias{examine.game}
\title{Badanie gry SuperFarmer}
\usage{
examine.game(namestrategy, rounds, die1, die2)
}
\arguments{
\item{namestrategy}{Nazwa wykonywanej strategii.}

\item{rounds}{Liczba wykonanych powtórzeń danej gry.}

\item{die1}{Kostka nr1.}

\item{die2}{Kostka nr2.}
}
\description{
Funkcja \code{examine.game} bada rozkład czasu gry SuperFarmer.
}
\examples{
examine.game(strategy_PDMS,100,die1,die2)

}

